+++ src/common/init.c
@@ -41,6 +41,8 @@ static void defaultConfig(void)
     #if defined(__linux__) || defined(__FreeBSD__)
     ffStrbufInit(&instance.config.playerName);
     ffStrbufInit(&instance.config.osFile);
+    #elif defined(_WIN32)
+    instance.config.wmiTimeout = 5000;
     #endif
 
     instance.config.showErrors = false;